refactor(dashboard): Move old directory file to new one
