+++ app/(route)/dashboard/saved/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata, NextPage } from 'next'
-import Favorite from '@/components/pages/dashboard/Favorite'
+import Favorite from '@/pages/dashboard/Favorite'
 
 export const metadata: Metadata = {
   title: 'Favorite'
